refactor(desktop): split four files back under the size ratchet
The send-path work pushed four files past the repository file-size gate.
Extract one cohesive unit from each rather than raising a ceiling:

- `runtime.rs` → `runtime/setup_payload.rs`: the readiness → payload JSON →
  `BUZZ_ACP_SETUP_PAYLOAD` write, as `apply_setup_payload_env`.
- `commands/agents.rs` → `commands/agents_create_fields.rs`: the pure
  create-request field validators/resolvers.
- `app_state.rs` → `app_state_accessors.rs`: the inherent `AppState`
  lock accessors, leaving the struct, builder, and identity resolution.
- `useMentionSendFlow.ts` → `useEnsureAgentMentionsReady.ts`: the mentioned
  agent access-policy / membership / detached-wake reconciliation.

Pure moves. The one behavioral difference is that the new hook depends on
the stable `mutateAsync` rather than the whole mutation object, so
`ensureManagedAgentMentionsReady` no longer changes identity every render.
